Update npx create-react-app path
If we directly install create-react-app in part1 folder, it becomes pretty difficult to ignore them when we commit exercises to Github.
It would be much cleaner with "npx create-react-app part1/practice"
Then, we can simply ignore the practice folder. All code along during the lecture can be done within this practice react app.